fix(1289-backup count): add key management key details

+++ src/db/include/awsmock/entity/kms/Key.h
@@ -118,7 +118,7 @@ namespace AwsMock::Database::Entity::KMS {
         /**
          * Pending window in days
          */
-        int pendingWindowInDays;
+        long pendingWindowInDays;
 
         /**
          * Scheduled deletion datetime
